pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["saet0sy <saet0sy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2.4"
markdown = "^3.4.4"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
